Moved spaceship building into its own launcher
